// libs/ng2-power-table/src/TableState/ITableState.interface.ts
import { EventEmitter } from "@angular/core";
import { SortOrder } from './../Sort/SortOrder.enum';

export interface ITableStateSearch {

}

export interface ITableStatePagination {
    start: number;
    end: number;
    pageSize: number;
    totalItemCount: number;
    numberOfPages: number;

    changed: EventEmitter<void>;
}

export interface ITableStateSort {
    predicate: string;
    order: SortOrder;

    changed: EventEmitter<void>;
}

export interface ITableState {
    sort: ITableStateSort;
    search: ITableStateSearch;
    pagination: ITableStatePagination;
}